Updated CMake to support ARM64 and x86_64 architectures on the new Apple M1 architecture.
If MUQ is compiled on OSX, we now check to see if arm instructions are supported.  If they are, the MUQ shared libraries will be built as universal binaries with support for both ARM and x86 instruction sets.  This requires universal binaries from all MUQ dependencies.  The internal build scripts have been updated to support this.